+++ lib/geo_pattern.rb
@@ -18,23 +18,23 @@ require 'geo_pattern/color_preset'
 
 require 'geo_pattern/generators/background_generator'
 
-require 'geo_pattern/pattern/base_pattern'
-require 'geo_pattern/pattern/chevron_pattern'
-require 'geo_pattern/pattern/concentric_circles_pattern'
-require 'geo_pattern/pattern/diamond_pattern'
-require 'geo_pattern/pattern/hexagon_pattern'
-require 'geo_pattern/pattern/mosaic_squares_pattern'
-require 'geo_pattern/pattern/nested_squares_pattern'
-require 'geo_pattern/pattern/octagon_pattern'
-require 'geo_pattern/pattern/overlapping_circles_pattern'
-require 'geo_pattern/pattern/overlapping_rings_pattern'
-require 'geo_pattern/pattern/plaid_pattern'
-require 'geo_pattern/pattern/plus_sign_pattern'
-require 'geo_pattern/pattern/sine_wave_pattern'
-require 'geo_pattern/pattern/square_pattern'
-require 'geo_pattern/pattern/tessellation_pattern'
-require 'geo_pattern/pattern/triangle_pattern'
-require 'geo_pattern/pattern/xes_pattern'
+require 'geo_pattern/structure_generators/base_generator'
+require 'geo_pattern/structure_generators/chevrons_generator'
+require 'geo_pattern/structure_generators/concentric_circles_generator'
+require 'geo_pattern/structure_generators/diamonds_generator'
+require 'geo_pattern/structure_generators/hexagons_generator'
+require 'geo_pattern/structure_generators/mosaic_squares_generator'
+require 'geo_pattern/structure_generators/nested_squares_generator'
+require 'geo_pattern/structure_generators/octagons_generator'
+require 'geo_pattern/structure_generators/overlapping_circles_generator'
+require 'geo_pattern/structure_generators/overlapping_rings_generator'
+require 'geo_pattern/structure_generators/plaid_generator'
+require 'geo_pattern/structure_generators/plus_signs_generator'
+require 'geo_pattern/structure_generators/sine_waves_generator'
+require 'geo_pattern/structure_generators/squares_generator'
+require 'geo_pattern/structure_generators/tessellation_generator'
+require 'geo_pattern/structure_generators/triangles_generator'
+require 'geo_pattern/structure_generators/xes_generator'
 
 require 'geo_pattern/pattern_generator'
